new flock-llm release
