+++ .gitignore
@@ -161,3 +161,4 @@ cython_debug/
 
 .DS_Store
 wandb/
+model/
